Removing KeyRunner.exitGame()
Most of the code which used it were logic error cases.  These cases are
now throwing exceptions.  The other cases were in the controller game
loop.  Due to the introduction of active rendering, it is now convenient
to just exit the game loop by setting the playModel state to QUIT.