Fixing background color of the Editor's Info Bar
This change sets the render color so that SDL_RenderClear explicitly
clears to black.  The normal play game mode was lucky in previous
commits that this color ended up being black.  While trouleshooting
editor mode, I found that the bar was getting clear to a dark gray
color--which was the background color of the buttons.