Controller Refactor
Controller is now an interface.  It has an abstract BaseController
implementation which is completed by PlayController and EditController.
These controllers are selected based on command line arguments from the
user.  The controllers have been simplified by requiring that the
rootLayer, model, renderer and window be passed to them.  This has made
it necessary to move the creation of these items into main

It also made since to put the homeless globals in main as well.  At
least they're all in one place now.

The end result of these refactor is that controller implementations are
dead simple and it's clear which behaviors are for editing and which are
for playing.