Adding Display and View to architecture
These will eventually completely replace Layers.  Display holds the
SDL_Window and SDL_Renderer.  It also holds a collection of View objects
which it draws onto the renderer.  Controller now relies on Display to
do this activity.

Controllers are responsible for building the View objects and passing
them to the Display (and potentially removing them later).

In order to minimize disruption on master, this refactor is being
managed on a branch.  This first commit contains the Display class, View
interface, and BaseView abstract class.  It also contains the first
Layer => View migration.  This was PlayInfoBarLayer => PlayInfoBarView.

The result is that very little displays right now, only the Play-mode
Info Bar.