Beginning to replace GridLayer with BoardView
This change reimplements GridLayer as a BoardView. The result is that
the game is surprisingly playable again. Editor is still not working
because it still uses GridLayer. Conveyors are functional, but they
don't animate.

Additional work is required to fully eliminate GridLayer. Particularly,
the EditController will need to be refactored to use BoardView. Also,
ConveyorAnimation needs some redesign. It is highly reliant on GridLayer
and I'm still not clear on where it falls in the MVC.