Views now retain their Animations
Display now advances each of its View's animations.  Controllers tell
the Display to do so.  After this change, it's become clear that
Displays should be part of the Controller, and not the View.

BoardView now stores its key and player animations in its standard
collection of Animations.  This allows them to be advanced automatically
by the controller.

Conveyors Animations are next.