PlayModel Grid Size now Non-Static
Previously, PlayModel::GRID_HEIGHT and GRID_WIDTH were static class
constants of PlayModel.  This change causes them to be returned from
member methods.  This will eventually allow the level to be fully
represented as member methods in the model.

GridLayer has mentions to these values hard coded to 16 and 25 until
GridLayer can be removed.  This is an acceptable compromise because
GridLayer is still heavily referenced, but is not part of any active
code.