Refactoring Conveyors
ConveyorAnimation has been removed. BoardView now manages the Animations
which represent conveyors. The new arrangement is much simpler, but will
rely on truly seamless conveyor animations in order to look right. I'm
not concerned about this for the moment as I've thought for a while that
the conveyor animations could look nicer.