Breaking down the Model
PlayModel and EditorModel now extend BaseModel which implements Model.
BaseModel provides a BoardModel which represents the board.  This makes
sense because both PlayModel and EditorModel need a BoardModel.

EditController now receives and EditorModel and PlayController continues
to receive a PlayModel.

BoardView is now restricted to just the BoardModel.

This change was required because it's now clear the the Controllers must
have a clear model hierarchy in order to deal with the complexities of
the editor.  Otherwise, we run the risk of mixing model, controller and
view elements when implementing GUI widgets.

The drawback of this refactor is that it now necessitates changes to
LevelManager which does not yet know about EditorModel.