Implementing ButtonView
Used ButtonLayer a starting point for implementing ButtonView. It's not
complete yet, but it's a start. In addition to the original functions of
ButtonLayer, ButtonView also receives a callback function for handling
clicks.

The EditorController was modified to merely exercise this function for
the exit button. This is not the final design, but I wanted something in
place to prove that the ButtonView was working correctly.

Next Display will need to be modified to determine which Views are being
clicked. Additionally, the click behaviors may need to move up to the
View interface.