Eliminating PlayInfoBarView
It has been fully replaced with LabelViews and RectangleViews.
LabelViews are now the parent class of ButtonViews.  Right now,
ButtonViews do little more than LabelViews, however this can change in
the future.

The sizing and appearance of labels and buttons are not quite as they
used to be or should be.  This is a byproduct of the fact that
LabelViews rely on automatic font sizing based on the dimensions of the
LabelView.

Also, because this code base does not yet support the concept of binding
data fields to the views, TimerView and LevelNumberView were introduced.
This subclasses of LabelView are KeyRunner-specific (just like
BoardView).  While model-field-to-view binding would be ideal, it's also
prohibitively difficult to implement and not worth it at this time.

That said, this system is consistent:  Views are required to have a
Model and they know how to get data out of the Model, format it and
present it.  Controllers create and control the Views and Models based
on user input and game logic.