View::onMouseDown(), onMouseUp()
Removed onClick() and replaced with onMouseDown() and onMouseUp(). This
is necessary because the editor will need this level of detail and also
because SDL does not generalize these events into 'onClick()' style
events anyway.

These callback functions now accept SDL_Events so that they have full
access to the information associated with the event.