Reorganizing Model Again
This change is a recognition that BaseModel is irrelevant.
EditController and PlayController are both represented by BoardModels,
but EditController requires some model elements that PlayController does
not (and vice versa). Therefore, it makes sense for there to be a
PlayBoardModel and EditorBoardModel. BaseModel, PlayModel and EditModel
have all been removed.