Tiles can be replaced by click-dragging
This change introduces the HoverBehavior enum class which the
EditorBoardModel uses to keep track what to do as the user clicks on and
hovers over the board. It currently supports two behaviors: POINT and
REPLACE. POINT represents the default behavior: display the highlighted,
selected tile type as the user hovers the mouse over the board. REPLACE
will actively replace tiles on the board with the selected tile type as
the user drags the mouse with the button down.