Editor Bugfixes
The EditorBoardView was never displaying the selected tile type on hover
due to a boolean reversal logic error. Additionally, dragging the player
or the key would cause the selected tile type to be 'stuck' in the spot
that the player or key used to be before dragging started.

This change fixes both of these issues.