TileType Warnings Addressed
TileType had a number of warnings. These warnings were addressed by
converting TileType into an enum class with 'int' as a base type. The
code which converts TileTypes into AnimationTypes now lives in
AnimationTypeFactory. This is a departure from Java-like enums.

Additionally, the enum names have changed to be a little more readable
(and shorter).