TitleScreen Play Button Starts Game
This was a relatively simple change after the main loop was modified to
react to state changes when controller game loops exit. Once they exit,
the model's state is checked. That state is used to determine which
model/controller duo to load next.

When title screen's controller exits in the model state "play", the play
controller and models are loaded.