Adding Visible Option Count to Menu View
Setting this to 0 will try to fit all options into the visible area of
the menu view. Otherwise, only a limited count of options will be fit.
Invisible options that are 'off screen', can still be accessed using
arrow keys. In the future, MenuViews will support scrolling in order to
allow these options to be accessible.