EditController's processInput() factored out
EditController now relies on BaseController's default input processing
function which delegates SDL Events to the appropriate view handler
(based either on what is currently in focus or on which view is under
the mouse--depending on the event).

Now EditController, PlayController and TitleScreenController all rely on
BaseController's processInput() function. The result is that
BaseController::processInput() can now be final because it is considered
correct and ideal for each view to be constructed with event handlers by
the controllers.