Adding MenuView::setOptionFontSize()
This method will set the font size of each of the internal buttons
within the menu view. The result is that each button will seem to have
the same size of text. '0' can still be used to automatically fit the
text within the button, however this should be used carefully on a
MenuView: if the text strings are of varied lengths, the options can
have irregularly sized text.