WIP: Player not bound to grid
This is a large set of changes required to unbind the player from the
grid. There are many more changes required, but this one gets many of
the pieces in the right spot. The game is not completely playable,
though, because player/item intersections are not yet working. This
means that most of the level ending mechanics are non-functional.

Also, the editor is not yet equipped to read and write levels in the new
format.

Much work to do, but this changes opens the door to smaller commits
henceforth.