Hard coding hitboxes for Key and Player
These are the two most important BoardEntity objects. With these hit
boxes hard coded, we can now experiment with game logic. Eventually, a
map of BoardEntityType -> hit box dimensions will need to exist so that
hard coding is not required.

It was decided that it was not good to write the hit box dimensions to
the level file. In the future, if hit box dimensions changed, this would
be a portability concern.