Fixing Two Bugs
(1). Hit Boxes were not being populated into BoardEntity objects
correctly because BoardEntity.getHitBoxes() returned a copy  of the
vector and not a reference.

(2). When picked up, the key was being redraw to pixel coordinates equal
to the tile coordinate width and height of the board (i.e. pixel 25,
19).  This change does away with the idea of moving the key off screen.
Instead, when the key is picked up by the player, it is simply not drawn
anymore.