Wall Hit Boxes Implemented
This change causes the LevelManager to create wall hit boxes for all
tiles of the wall type. The list of wall hit boxes are cleared before a
new level is loaded.

This change has created a temporary problem, though: movement is capped
to 1 pixel per frame. This is required for the game to be basically
playable because moving 4 pixel per frame was making so that the player
couldn't move through 1 tile wide gaps between walls. The big problem
here is that 1 pixel per frame is VERY SLOW and renders the game
un-winnable.