Porting Levels to JSON
The challenges encountered with making some simple changes to the binary
level file format have prompted me to port the level files to JSON. JSON
based files should make it considerably easier to add or remove fields
to the level files in the future--which is something I expected will be
necessary in order to add compelling new features.

This change retains the older BinaryLevelManager so that it can be
checked into Git's history. A subsequent commit will remove it.